+++ pkg/util/admission/BUILD.bazel
@@ -4,8 +4,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "admission",
     srcs = [
+        "admission.go",
         "disk_bandwidth.go",
-        "doc.go",
         "grant_coordinator.go",
         "granter.go",
         "io_load_listener.go",
